중간통합(한별)0614

+++ carini/carini/src/main/java/com/car/service/ModelService.java
@@ -13,7 +13,7 @@ public interface ModelService {
 	
 	long getTotalRowCount(Car car);
 	List<Car> getAllCarList();
-	Car getCarbyId(Long carId);
+	Car getCarbyId(int carId);
 	Page<Car> filterCars(Pageable pageable, Long carMinPrice, Long carMaxPrice, String carSize, String carFuel);
